+++ drizzle/0020_lively_ser_duncan.sql
@@ -1,0 +1,4 @@
+ALTER TABLE `stem` ADD `midi_url` text;--> statement-breakpoint
+ALTER TABLE `stem` ADD `midi_pathname` text;--> statement-breakpoint
+ALTER TABLE `stem` ADD `midi_filename` text;--> statement-breakpoint
+ALTER TABLE `stem` ADD `midi_size_bytes` integer;